--- algorithms/main.go
@@ -1,16 +1,0 @@
-/*
-	The examples algorithms found in here have been obtained from:
-	https://medium.com/@uzairahmed01/algorithm-and-data-structure-in-golang-2869da82723e
-*/
-
-package main
-
-import (
-	"fmt"
-)
-
-func main() {
-	fmt.Println(">> SORTING ALGORITHMS")
-
-	sorting_algo()
-}
